requirements.txt
flask>=2.3.0
pyjwt>=2.7.0
cryptography>=41.0.0
pyyaml>=6.0
requests>=2.31.0
